Move temperature related plugins into a temperature sub-directory. Move related tests similarly, and update all imports. (#2119)
