+++ improver/api/__init__.py
@@ -22,7 +22,7 @@ PROCESSING_MODULES = {
     "ApplyDecisionTree": "improver.categorical.decision_tree",
     "ApplyDzRescaling": "improver.calibration.dz_rescaling",
     "ApplyEMOS": "improver.calibration.ensemble_calibration",
-    "ApplyGriddedLapseRate": "improver.lapse_rate",
+    "ApplyGriddedLapseRate": "improver.temperature.lapse_rate",
     "ApplyNeighbourhoodProcessingWithAMask": "improver.nbhood.use_nbhood",
     "ApplyOrographicEnhancement": "improver.nowcasting.utilities",
     "ApplyRainForestsCalibration": "improver.calibration.rainforest_calibration",
@@ -73,7 +73,7 @@ PROCESSING_MODULES = {
     "HumidityMixingRatio": "improver.psychrometric_calculations.psychrometric_calculations",
     "Integration": "improver.utilities.mathematical_operations",
     "InterpolateUsingDifference": "improver.utilities.interpolation",
-    "LapseRate": "improver.lapse_rate",
+    "LapseRate": "improver.temperature.lapse_rate",
     "LightningFromCapePrecip": "improver.lightning",
     "LightningMultivariateProbability_USAF2024": "improver.lightning",
     "ManipulateReliabilityTable": "improver.calibration.reliability_calibration",
@@ -120,7 +120,7 @@ PROCESSING_MODULES = {
     "Threshold": "improver.threshold",
     "TriangularWeightedBlendAcrossAdjacentPoints": "improver.blending.blend_across_adjacent_points",
     "VerticalUpdraught": "improver.wind_calculations.vertical_updraught",
-    "VirtualTemperature": "improver.virtual_temperature",
+    "VirtualTemperature": "improver.temperature.virtual_temperature",
     "VisibilityCombineCloudBase": "improver.visibility.visibility_combine_cloud_base",
     "WeightAndBlend": "improver.blending.calculate_weights_and_blend",
     "WeightedBlendAcrossWholeDimension": "improver.blending.weighted_blend",
