+++ improver_tests/temperature/lapse_rate/test_ApplyGriddedLapseRate.py
@@ -11,11 +11,11 @@ import numpy as np
 from iris.tests import IrisTest
 
 from improver.constants import DALR
-from improver.lapse_rate import ApplyGriddedLapseRate
 from improver.synthetic_data.set_up_test_cubes import (
     add_coordinate,
     set_up_variable_cube,
 )
+from improver.temperature.lapse_rate import ApplyGriddedLapseRate
 
 
 class Test_process(IrisTest):
